requirements.txt
django==5.0.6
daphne==4.1.2
channels==4.1.0
channels-redis==4.2.0
psycopg2==2.9.9
djangorestframework==3.15.1
djangorestframework-jwt==1.11.0
djangorestframework_simplejwt==5.3.1
django-cors-headers==4.3.1
shortuuid==1.0.13